Refactored mappers and their tests
