+++ plugins/azure-devops-backend/src/api/AzureDevOpsApi.ts
@@ -19,9 +19,7 @@ import {
   BuildDefinitionReference,
 } from 'azure-devops-node-api/interfaces/BuildInterfaces';
 import {
-  BuildResult,
   BuildRun,
-  BuildStatus,
   DashboardPullRequest,
   GitTag,
   Policy,
@@ -65,6 +63,12 @@ import {
   DefaultAzureDevOpsCredentialsProvider,
   ScmIntegrations,
 } from '@backstage/integration';
+import {
+  mappedBuildRun,
+  mappedGitTag,
+  mappedPullRequest,
+  mappedRepoBuild,
+} from './mappers';
 
 /** @public */
 export class AzureDevOpsApi {
@@ -493,76 +497,4 @@ export class AzureDevOpsApi {
     );
     return { url, content };
   }
-}
-
-export function mappedRepoBuild(build: Build): RepoBuild {
-  return {
-    id: build.id,
-    title: [build.definition?.name, build.buildNumber]
-      .filter(Boolean)
-      .join(' - '),
-    link: build._links?.web.href ?? '',
-    status: build.status ?? BuildStatus.None,
-    result: build.result ?? BuildResult.None,
-    queueTime: build.queueTime?.toISOString(),
-    startTime: build.startTime?.toISOString(),
-    finishTime: build.finishTime?.toISOString(),
-    source: `${build.sourceBranch} (${build.sourceVersion?.slice(0, 8)})`,
-    uniqueName: build.requestedFor?.uniqueName ?? 'N/A',
-  };
-}
-
-export function mappedGitTag(
-  gitRef: GitRef,
-  linkBaseUrl: string,
-  commitBaseUrl: string,
-): GitTag {
-  return {
-    objectId: gitRef.objectId,
-    peeledObjectId: gitRef.peeledObjectId,
-    name: gitRef.name?.replace('refs/tags/', ''),
-    createdBy: gitRef.creator?.displayName ?? 'N/A',
-    link: `${linkBaseUrl}${encodeURIComponent(
-      gitRef.name?.replace('refs/tags/', '') ?? '',
-    )}`,
-    commitLink: `${commitBaseUrl}/${encodeURIComponent(
-      gitRef.peeledObjectId ?? '',
-    )}`,
-  };
-}
-
-export function mappedPullRequest(
-  pullRequest: GitPullRequest,
-  linkBaseUrl: string,
-): PullRequest {
-  return {
-    pullRequestId: pullRequest.pullRequestId,
-    repoName: pullRequest.repository?.name,
-    title: pullRequest.title,
-    uniqueName: pullRequest.createdBy?.uniqueName ?? 'N/A',
-    createdBy: pullRequest.createdBy?.displayName ?? 'N/A',
-    creationDate: pullRequest.creationDate?.toISOString(),
-    sourceRefName: pullRequest.sourceRefName,
-    targetRefName: pullRequest.targetRefName,
-    status: pullRequest.status,
-    isDraft: pullRequest.isDraft,
-    link: `${linkBaseUrl}/${pullRequest.pullRequestId}`,
-  };
-}
-
-export function mappedBuildRun(build: Build): BuildRun {
-  return {
-    id: build.id,
-    title: [build.definition?.name, build.buildNumber]
-      .filter(Boolean)
-      .join(' - '),
-    link: build._links?.web.href ?? '',
-    status: build.status ?? BuildStatus.None,
-    result: build.result ?? BuildResult.None,
-    queueTime: build.queueTime?.toISOString(),
-    startTime: build.startTime?.toISOString(),
-    finishTime: build.finishTime?.toISOString(),
-    source: `${build.sourceBranch} (${build.sourceVersion?.slice(0, 8)})`,
-    uniqueName: build.requestedFor?.uniqueName ?? 'N/A',
-  };
 }
